disease_prediction_app/requirements.txt

mysql-connector-python
